+++ packages/liveblocks-core/src/notifications.ts
@@ -26,11 +26,6 @@ import type {
 
 const MARK_INBOX_NOTIFICATIONS_AS_READ_BATCH_DELAY = 50;
 
-export type GetInboxNotificationsOptions = {
-  limit?: number;
-  since?: Date;
-};
-
 export function createNotificationsApi<M extends BaseMetadata>({
   baseUrl,
   authManager,
@@ -102,7 +97,27 @@ export function createNotificationsApi<M extends BaseMetadata>({
     return body;
   }
 
-  async function getInboxNotifications(options?: GetInboxNotificationsOptions) {
+  async function getInboxNotifications() {
+    const json = await fetchJson<{
+      threads: ThreadDataPlain<M>[];
+      inboxNotifications: InboxNotificationDataPlain[];
+      deletedThreads: ThreadDeleteInfoPlain[];
+      deletedInboxNotifications: InboxNotificationDeleteInfoPlain[];
+      meta: {
+        requestedAt: string;
+      };
+    }>("/inbox-notifications", undefined, {});
+
+    return {
+      threads: json.threads.map(convertToThreadData),
+      inboxNotifications: json.inboxNotifications.map(
+        convertToInboxNotificationData
+      ),
+      requestedAt: new Date(json.meta.requestedAt),
+    };
+  }
+
+  async function getInboxNotificationsSince(options: { since: Date }) {
     const json = await fetchJson<{
       threads: ThreadDataPlain<M>[];
       inboxNotifications: InboxNotificationDataPlain[];
@@ -112,24 +127,21 @@ export function createNotificationsApi<M extends BaseMetadata>({
         requestedAt: string;
       };
     }>("/inbox-notifications", undefined, {
-      limit: options?.limit,
-      since: options?.since?.toISOString(),
+      since: options.since.toISOString(),
     });
 
     return {
-      threads: json.threads.map((thread) => convertToThreadData(thread)),
-      inboxNotifications: json.inboxNotifications.map((notification) =>
-        convertToInboxNotificationData(notification)
-      ),
-      deletedThreads: json.deletedThreads.map((info) =>
-        convertToThreadDeleteInfo(info)
-      ),
-      deletedInboxNotifications: json.deletedInboxNotifications.map((info) =>
-        convertToInboxNotificationDeleteInfo(info)
-      ),
-      meta: {
-        requestedAt: new Date(json.meta.requestedAt),
+      threads: {
+        updated: json.threads.map(convertToThreadData),
+        deleted: json.deletedThreads.map(convertToThreadDeleteInfo),
       },
+      inboxNotifications: {
+        updated: json.inboxNotifications.map(convertToInboxNotificationData),
+        deleted: json.deletedInboxNotifications.map(
+          convertToInboxNotificationDeleteInfo
+        ),
+      },
+      requestedAt: new Date(json.meta.requestedAt),
     };
   }
 
@@ -193,6 +205,7 @@ export function createNotificationsApi<M extends BaseMetadata>({
 
   return {
     getInboxNotifications,
+    getInboxNotificationsSince,
     getUnreadInboxNotificationsCount,
     markAllInboxNotificationsAsRead,
     markInboxNotificationAsRead,