+++ packages/liveblocks-react/src/__tests__/useCreateComment.test.tsx
@@ -1,8 +1,8 @@
 import type { CommentBody } from "@liveblocks/core";
+import { nanoid } from "@liveblocks/core";
 import { act, renderHook, waitFor } from "@testing-library/react";
 import { addMinutes } from "date-fns";
 import { setupServer } from "msw/node";
-import { nanoid } from "nanoid";
 import React from "react";
 
 import {